requirements.txt
streamlit==1.28.1
streamlit-option-menu==0.3.6
pdfplumber==0.10.3
python-docx==0.8.11
PyMuPDF==1.23.8
textract==1.6.3
spacy==3.7.2
sentence-transformers==2.2.2
rapidfuzz==3.5.2
scikit-learn==1.3.2
hdbscan==0.8.30
numpy==1.24.3
pandas==2.0.3
plotly==5.17.0
python-dotenv==1.0.0
openai==1.3.9
google-generativeai==0.3.0
anthropic==0.7.10
requests==2.31.0
sqlalchemy==2.0.23
supabase==2.0.0
pytest==7.4.3
python-multipart==0.0.6